requirements.txt
attrs==25.1.0
certifi==2025.1.31
h11==0.14.0
idna==3.10
numpy==2.2.3
outcome==1.3.0.post0
pandas==2.2.3
pyodbc==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
